Use GNU-style formatting directives on Windows
By default, on Windows, GCC will check to make sure that formatting directives
that are not supported on Windows, like %llu and %zu, are not used.  Recent
versions of MinGW ship POSIX-compliant printf replacements that do support
these directives, but GCC is not smart enough to realize this on its own.